lib conncache bump var to size_t

+++ lib/conncache.c
@@ -100,7 +100,7 @@ static void free_bundle_hash_entry(void *freethis)
   bundle_destroy(b);
 }
 
-int Curl_conncache_init(struct conncache *connc, int size)
+int Curl_conncache_init(struct conncache *connc, size_t size)
 {
   /* allocate a new easy handle to use when closing cached connections */
   connc->closure_handle = curl_easy_init();
@@ -108,7 +108,7 @@ int Curl_conncache_init(struct conncache *connc, int size)
     return 1; /* bad */
   connc->closure_handle->state.internal = true;
 
-  Curl_hash_init(&connc->hash, (size_t)size, Curl_hash_str,
+  Curl_hash_init(&connc->hash, size, Curl_hash_str,
                  Curl_str_key_compare, free_bundle_hash_entry);
   connc->closure_handle->state.conn_cache = connc;
